[fix] 메인화면 설정 수정, ui 수정
