+++ Assets/Scenes/StartScene.unity
@@ -282,10 +282,10 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 248280547}
-  - component: {fileID: 248280549}
   - component: {fileID: 248280548}
   - component: {fileID: 248280550}
   - component: {fileID: 248280551}
+  - component: {fileID: 248280552}
   m_Layer: 0
   m_Name: GameObject
   m_TagString: Untagged
@@ -320,21 +320,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b460298dcdc911f44b9da2a38b562dfd, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &248280549
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 248280546}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ff7cddae66552ef43b341a1245bf28d3, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  settingsPanel: {fileID: 514824661}
-  pauseButton: {fileID: 720187486}
-  resumeButton: {fileID: 1287305294}
 --- !u!82 &248280550
 AudioSource:
   m_ObjectHideFlags: 0
@@ -445,6 +430,19 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   masterMixer: {fileID: 24100000, guid: d35149c666304814bba2eb34f5a30cd7, type: 2}
   volumeSlider: {fileID: 821194187}
+--- !u!114 &248280552
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 248280546}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 731fc217494b17c4bbfca58c7d83eec2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  settingsPanel: {fileID: 514824661}
 --- !u!1 &288930198
 GameObject:
   m_ObjectHideFlags: 0
@@ -1455,9 +1453,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: SceneChanger, Assembly-CSharp
-        m_MethodName: GoMainScene
+      - m_Target: {fileID: 248280552}
+        m_TargetAssemblyTypeName: MapManager, Assembly-CSharp
+        m_MethodName: PauseGame
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -2345,7 +2343,19 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 1287305295}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls: []
+      m_Calls:
+      - m_Target: {fileID: 248280552}
+        m_TargetAssemblyTypeName: MapManager, Assembly-CSharp
+        m_MethodName: ResumeGame
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
 --- !u!114 &1287305295
 MonoBehaviour:
   m_ObjectHideFlags: 0
