+++ Assets/Scenes/TestScenes/GamePlayTest.unity
@@ -600,7 +600,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -9.18, y: -19.9}
+  m_AnchoredPosition: {x: -9.18, y: -20}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &144495336
@@ -675,7 +675,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0.5192337, y: -96.41561}
+  m_AnchoredPosition: {x: 0.5192337, y: -87}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &207366147
@@ -1483,7 +1483,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 22.86544, y: -96.41561}
+  m_AnchoredPosition: {x: 22.86544, y: -87}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &717233127
@@ -1833,7 +1833,7 @@ RectTransform:
   m_GameObject: {fileID: 816384616}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 2.7040267}
-  m_LocalScale: {x: 2.5, y: 1, z: 1}
+  m_LocalScale: {x: 2.5, y: 0.93, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 2052824097}
@@ -1844,7 +1844,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -63.987183, y: -69.07953}
+  m_AnchoredPosition: {x: -63.987183, y: -66.3}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &816384619
@@ -1859,7 +1859,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 2100000, guid: 9409b3ce8a4ab1e48a87ea4d869da443, type: 2}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_RaycastTarget: 1
   m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
@@ -1867,7 +1867,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: -856218377, guid: 28f7744715740cb489b6d795f1ca1013, type: 3}
+  m_Sprite: {fileID: 273909420, guid: 7df2f3847b326af40936530d011932e3, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -2409,7 +2409,7 @@ RectTransform:
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 50, y: 50}
+  m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1027015346
 MonoBehaviour:
@@ -3003,7 +3003,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 1, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 0, y: 0}
+  m_AnchoredPosition: {x: -15, y: -15}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 1, y: 1}
 --- !u!1 &1291869543
@@ -3043,7 +3043,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -4, y: -42.7}
+  m_AnchoredPosition: {x: -4, y: -43.6}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1291869546
@@ -3058,7 +3058,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 2100000, guid: 9409b3ce8a4ab1e48a87ea4d869da443, type: 2}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_RaycastTarget: 1
   m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
@@ -3066,7 +3066,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 423692872, guid: 28f7744715740cb489b6d795f1ca1013, type: 3}
+  m_Sprite: {fileID: 1064403804, guid: 7df2f3847b326af40936530d011932e3, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -3862,7 +3862,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -20.845009, y: -19.9}
+  m_AnchoredPosition: {x: -19.7, y: -19.9}
   m_SizeDelta: {x: 34.83, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1664426420
@@ -3937,7 +3937,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -20.8, y: -19.9}
+  m_AnchoredPosition: {x: -20.8, y: -20}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1720864074
@@ -4155,6 +4155,7 @@ GameObject:
   - component: {fileID: 1765772168}
   - component: {fileID: 1765772167}
   - component: {fileID: 1765772169}
+  - component: {fileID: 1765772170}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: Untagged
@@ -4206,8 +4207,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   settingsPanel: {fileID: 1446266540}
+--- !u!114 &1765772170
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1765772166}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7180818b514e79d46a930a5a2180100a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   backgroundPrefab: {fileID: 6015415691774125930, guid: b9252a638cdd650459a44665e94bddad, type: 3}
-  playerTransform: {fileID: 9207805186883838742, guid: b9252a638cdd650459a44665e94bddad, type: 3}
+  playerTransform: {fileID: 405517137}
   maxPieces: 3
   destroyPointX: -110
 --- !u!1 &1914322861
@@ -4496,7 +4509,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -22, y: -96.41561}
+  m_AnchoredPosition: {x: -22, y: -87}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1960741290
@@ -4975,7 +4988,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -32.51, y: -19.9}
+  m_AnchoredPosition: {x: -32.51, y: -20}
   m_SizeDelta: {x: 11.5, y: 32.3}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &2052824098
